requirements.txt
altgraph==0.17.4
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
astroid==3.3.11
asttokens==3.0.0
attrs==25.3.0
branca==0.8.1
Brotli==1.1.0
cachetools==5.5.2
Cartopy==0.24.1
certifi==2025.8.3
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.1
coverage==7.10.5
cryptography==45.0.7
cssselect2==0.8.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.16
decorator==5.2.1
dill==0.4.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.2
folium==0.19.5
fonttools==4.56.0
future==1.0.0
geocoder==1.38.1
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
nbformat==5.10.4
nbval==0.11.0
nest-asyncio==1.6.0
netCDF4==1.7.2
numpy==2.2.3
oauthlib==3.3.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
parso==0.8.5
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pygmt==0.14.2
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pylint==3.3.8
pyparsing==3.2.1
pypdf==6.0.0
pyperclip==1.9.0
pyphen==0.17.2
pyproj==3.7.1
pyshp==2.3.1
pyspnego==0.11.2
pytest==8.4.1
pytest-cov==6.2.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.2
ratelim==0.1.6
referencing==0.36.2
reportlab==4.4.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests_ntlm==1.3.0
rpds-py==0.27.0
rsa==4.9.1
scipy==1.16.1
setuptools==80.9.0
shapely==2.0.7
SharePlum==0.5.1
six==1.17.0
sqlparse==0.5.3
sspilib==0.3.0
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tk==0.1.0
tkintermapview==1.29
tomlkit==0.13.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
xarray==2025.1.2
xhtml2pdf==0.2.17
xyzservices==2025.1.0
zopfli==0.2.3.post1
